refactor(tests): split integration.rs into focused test modules (#355)
Split the 4566-line integration.rs monolith into 8 focused test binaries:

- tests/broker.rs     — core UDS broker protocol (23 tests)
- tests/oneshot.rs    — oneshot send/poll/pull/who commands (17 tests)
- tests/auth.rs       — token auth and admin commands (17 tests)
- tests/ws.rs         — WebSocket protocol and REST API (15 tests)
- tests/daemon.rs     — multi-room daemon and DM enforcement (15 tests)
- tests/scripted.rs   — multi-agent coordination scenarios (8 tests)
- tests/rest_query.rs — REST /api/{room}/query endpoint (11 tests)
- tests/room_lifecycle.rs — UDS CREATE:/DESTROY: and REST room creation (21 tests)

Moved all shared fixtures into tests/common/mod.rs. No tests removed;
total integration test count unchanged at 127.

Closes #355
